[FEAT] 여행경로 연동 및 날짜 수정 테스트 #27
